% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avdata-methods.R
\name{avdata-methods}
\alias{avdata-methods}
\alias{avdata}
\alias{avdata_import}
\alias{avdata,azure-method}
\alias{avdata_import,azure-method}
\title{Azure "Reference" and "Other" Data methods on AnVIL}
\usage{
\S4method{avdata}{azure}(
  namespace = avworkspace_namespace(),
  name = avworkspace_name(),
  ...,
  platform = cloud_platform()
)

\S4method{avdata_import}{azure}(
  .data,
  namespace = avworkspace_namespace(),
  name = avworkspace_name(),
  ...,
  platform = cloud_platform()
)
}
\arguments{
\item{...}{Additional arguments passed to lower level functions (not used)}

\item{platform}{\code{azure()} The cloud platform class to dispatch on as given by
\link[AnVILBase:cloud_platform]{AnVILBase::cloud_platform}. Typically not set manually as
\code{cloud_platform()} returns the \code{"azure"} class for Azure workspaces on
AnVIL.}
}
\description{
This file contains methods for working with "Reference" and
"Other" data on AnVIL.
}
\section{Functions}{
\itemize{
\item \code{avdata(azure)}: List the available "Reference" and "Other" data

\item \code{avdata_import(azure)}: Import "Reference" and "Other" data to an AnVIL
workspace

}}
\examples{
if (az_exists() && identical(get_platform(), "AnVILAz")) {
   ## from within AnVIL
   data <- avdata()
   avdata_import(data)
}

}
