% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avnotebooks-methods.R
\name{avnotebooks-methods}
\alias{avnotebooks-methods}
\alias{avnotebooks}
\alias{avnotebooks_localize}
\alias{avnotebooks_delocalize}
\alias{avnotebooks,azure-method}
\title{Azure Notebook Management}
\usage{
\S4method{avnotebooks}{azure}(
  namespace = avworkspace_namespace(),
  name = avworkspace_name(),
  local = FALSE,
  ...,
  platform = cloud_platform()
)
}
\arguments{
\item{namespace}{\code{character(1)} AnVIL workspace namespace as returned by,
e.g., \code{avworkspace_namespace()}}

\item{name}{\code{character(1)} AnVIL workspace name as returned by, eg.,
\code{avworkspace_name()}.}

\item{local}{\code{logical(1)} notebooks located on the workspace
(\code{local = FALSE}, default) or runtime / local instance (\code{local = TRUE}).
When \code{local = TRUE}, the notebook path is
\verb{<workspace_data_service_url()>/analyses}.}

\item{...}{Additional arguments passed to lower level functions (not used)}

\item{platform}{\code{azure()} The cloud platform class to dispatch on as given by
\link[AnVILBase:cloud_platform]{AnVILBase::cloud_platform}. Typically not set manually as
\code{cloud_platform()} returns the \code{"azure"} class for Azure workspaces on
AnVIL.}
}
\value{
\code{avnotebooks()} returns a character vector of files located
in the workspace 'analyses/' folder path, or on the local file
system.
}
\description{
\code{avnotebooks()} lists the notebooks in the current workspace.
}
\section{Functions}{
\itemize{
\item \code{avnotebooks(azure)}: List the notebooks in the current workspace

}}
\examples{
library(AnVILBase)
if (
    az_exists() && identical(avplatform_namespace(), "AnVILAz") &&
    nzchar(avworkspace_name())
)
    avnotebooks()

}
