% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az.R
\name{az-utilities}
\alias{az-utilities}
\alias{az_exists}
\title{az health check helpers}
\usage{
az_exists()
}
\value{
az_exists returns a logical value indicating the presence of the \code{az}
command line utility. \code{az_healthcheck} returns a logical value indicating
the presence of the \code{az} command line utility and the essential environment
variables.
}
\description{
These functions provide checks for essential workspace tools and
variables. \code{az_exists} checks for the presence of the \code{az} command line
utility. \code{az_healthcheck} checks for the presence of the \code{az} command line
as well as the essential environment variables.
}
\examples{
if (interactive()) {
    az_exists()
    az_healthcheck()
}
}
