% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az_copy.R
\name{az_copy}
\alias{az_copy}
\alias{az_copy_from_storage}
\alias{az_copy_list}
\alias{az_copy_copy}
\alias{az_copy_rm}
\alias{az_copy_to_storage}
\title{Azure Copy command line utility interface}
\usage{
az_copy_from_storage(from, to = ".")

az_copy_to_storage(from, to)

az_copy_list()

az_copy_rm(blob_file)
}
\arguments{
\item{from}{\code{character(1)} A relative file path corresponding to either the
remote (\code{az_copy_from_storage}) or local (\code{az_copy_to_storage}) file
location. Remote locations should be relative to the base directory in the
Azure Storage Container e.g., \code{analyses/jupyter.log}.}

\item{to}{\code{character(1)} A relative file path corresponding to either the
remote (\code{az_copy_to_storage}) or local (\code{az_copy_from_storage}) file
location. Remote locations should be relative to the base directory in the
Azure Storage Container. When not specified, it will default to the base
directory of the remote location. The \code{to} path can be a folder path but
must end in a forward slash (\code{/}). If the \code{to} path points to a
non-existent directory, it will be created.}

\item{blob_file}{\code{character(1)} A relative path to a file in the Azure
Storage Container to be removed}
}
\value{
\itemize{
\item \code{az_copy_list} - a \code{tibble} of files and metadata
\item \code{az_copy_from_storage} - called for the side effect of copying a file
\strong{from} the Azure Storage Container
\item \code{az_copy_to_storage} - called for the side effect of copying a file \strong{to}
the Azure Storage Container
\item \code{az_copy_rm} - called for the side effect of removing a file
}
}
\description{
These functions invoke the \code{azcopy} command line utility. The
utilities make use of a managed SAS token to mainly transfer files from the
Azure workspace to the Azure Storage container. See \code{get_sas_token} for
credential details.
}
\details{
\itemize{
\item \code{az_copy_list} - list all the files in the Azure Storage Container
\item \code{az_copy_copy} - copy a file from the Azure workspace to the container
\item \code{az_copy_rm} - remove a file from the Azure Storage Container
}
}
\examples{
if (interactive()) {

  az_copy_list()

  ## local -> remote
  az_copy_to_storage("jupyter.log", "analyses/jupyter.log")
  az_copy_to_storage("jupyter.log", "analyses/test/")
  ## placed in the base storage UUID directory
  az_copy_to_storage("jupyter.log")

  ## remote -> local
  az_copy_from_storage("analyses/jupyter.log", "jupyter.log")
  ## should be an error
  az_copy_from_storage("analyses/", "jupyter.log")
  ## download to the current directory
  az_copy_from_storage("analyses/jupyter.log")

  az_copy_rm("analyses/jupyter.log")

}
}
