% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{az_token}
\alias{az_token}
\title{Generate an Azure authentication token with the \code{az} command line
utility}
\usage{
az_token()
}
\value{
A character string containing the authentication token with a
"Bearer" prefix.
}
\description{
This function generates an Azure authentication token with the
\code{az} command line utility. The token is used to authenticate with the Azure
services. This function is called internally by the \verb{az_*} functions. It
is not meant to be called directly.
}
\keyword{internal}
