% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sas_token.R
\name{get_sas_token}
\alias{get_sas_token}
\title{Obtain the Shared Access Signature (SAS) for the Azure Storage}
\usage{
get_sas_token(as = "parsed", sasExpirationDuration = 28800)
}
\arguments{
\item{as}{desired type of output: \code{raw}, \code{text} or
\code{parsed}. \code{content} attempts to automatically figure out
which one is most appropriate, based on the content-type.}

\item{sasExpirationDuration}{\code{numeric(1)} The number of seconds until the SAS
token expires (default: 28,800 seconds)}
}
\value{
A list of two elements named \code{token} and \code{url}
}
\description{
The function provides a user delegation SAS token for management of
resources. Mainly used in other functions to move files to and from the
Azure Storage Container
}
\examples{
if (interactive()) {
  sas <- get_sas_token()
  sas[["token"]]
  sas[["url"]]
}
}
