% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_tsv.R
\name{upload_tsv}
\alias{upload_tsv}
\title{Upload a TSV file to WDS URL}
\usage{
upload_tsv(
  tsv_file,
  type = tools:::file_path_sans_ext(basename(tsv_file)),
  version = "v0.2"
)
}
\arguments{
\item{tsv_file}{\code{character(1)} A path to a tab-separated values file}

\item{type}{\code{character(1)} A nickname for the uploaded dataset important for
retreival. By default, the file name will be used.}

\item{version}{\code{character(1)} The version of the API. Set to \code{v0.2} by
default.}
}
\value{
The contents of the API POST request after uploading the TSV file
}
\description{
A function to move a flat Tab-Separated Values (TSV) file into the
Azure workspace using the Workspace Data Services (WDS) API
}
\examples{
if (interactive()) {
  type <- "model"
  mtcars_tbl <-
      mtcars |>
      as_tibble(rownames = "model_id") |>
      mutate(model_id = gsub(" ", "-", model_id))

  tsv_file <- tempfile()
  readr::write_tsv(mtcars_tbl, tsv_file)
  upload_tsv(
    tsv_file = tsvfile,
    type = "testData",
  )
}
}
