% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_tsv.R
\name{upload_tsv}
\alias{upload_tsv}
\alias{retrieve_tsv}
\alias{delete_tsv_row}
\alias{delete_tsv}
\alias{workspace-data-ops}
\title{Functions to work with workspace data}
\usage{
upload_tsv(
  tsv_file,
  type = tools:::file_path_sans_ext(basename(tsv_file)),
  primaryKey = NULL,
  api_version = .WDS_API_VERSION
)

retrieve_tsv(type, api_version = .WDS_API_VERSION)

delete_tsv_row(type, id, api_version = .WDS_API_VERSION)

delete_tsv(type, api_version = .WDS_API_VERSION)
}
\arguments{
\item{tsv_file}{\code{character(1)} A path to a tab-separated values file}

\item{type}{\code{character(1)} A nickname for the uploaded dataset important for
retreival. By default, the file name will be used.}

\item{primaryKey}{\code{character(1)} The optional column name to uniquely
identify a record. By default, the first column is used as the primary
key and all values in the column must be unique.}

\item{api_version}{\code{character(1)} The version of the Workspace Data Service
API. Set to the value of the internal \code{.WDS_API_VERSION} variable by
default. See the current version with \code{AnVILAz:::.WDS_API_VERSION}.}

\item{id}{\code{character(1)} The value in the \code{primaryKey} column that
indicates the row to be removed.}
}
\value{
\itemize{
\item \code{upload_tsv} - A response list indicating successful upload
\item \code{retrieve_tsv} - A \code{tibble} corresponding to the data labeled with \code{type}
\item \code{delete_tsv_row}; \code{delete_tsv} - When successful, a \code{NULL} value
}
}
\description{
These group of functions will allow you to manipulate data in
the "DATA" tab. Example operations include moving a flat Tab-Separated
Values (TSV) file into the workspace, deleting records, deleting tables,
and retrieving the data table.
}
\details{
These functions use the Workspace Data Services (WDS) API. Current
operations that affect the "DATA" tab include:
\itemize{
\item \code{upload_tsv} - a \code{POST} request using a TSV file that populates the data
\item \code{retrieve_tsv} - a \code{GET} request with the data name (\code{type} argument) in
\code{upload_tsv} to download the data
\item \code{delete_tsv_row} - a \code{DELETE} request to remove a record or row from
\code{type}
\item \code{delete_tsv} - a \code{DELETE} request to remove then entire data set (\code{type})
}
}
\examples{
if (interactive()) {
  library(dplyr)
  type <- "model"
  mtcars_tbl <-
      mtcars |>
      as_tibble(rownames = "model_id") |>
      mutate(model_id = gsub(" ", "-", model_id))

  tsv_file <- tempfile()
  readr::write_tsv(mtcars_tbl, tsv_file)
  upload_tsv(
    tsv_file = tsvfile,
    type = "testData",
    primaryKey = "model_id"
  )
}
}
