% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnVILAz-package.R
\docType{package}
\name{AnVILAz-package}
\alias{AnVILAz}
\alias{AnVILAz-package}
\title{AnVILAz: R / Bioconductor Support for the AnVIL Azure Platform}
\description{
The AnVIL is a cloud computing resource developed in part by the National Human Genome Research Institute. The AnVILAz package supports end-users and developers using the AnVIL platform in the Azure cloud.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Bioconductor/AnVILAz}
  \item Report bugs at \url{https://github.com/Bioconductor/AnVILAz/issues}
}

}
\author{
\strong{Maintainer}: Marcel Ramos \email{marcel.ramosperez@roswellpark.org} (\href{https://orcid.org/0000-0002-3242-0582}{ORCID})

Authors:
\itemize{
  \item Martin Morgan \email{mtmorgan.bioc@gmail.com} (\href{https://orcid.org/0000-0002-5874-8148}{ORCID}) [contributor]
}

}
\keyword{internal}
