% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avtable-methods.R
\name{avtable-methods}
\alias{avtable-methods}
\alias{avtable,azure-method}
\alias{avtables,azure-method}
\alias{avtable_import_set,azure-method}
\title{AnVIL Azure table ("type") methods}
\usage{
\S4method{avtable}{azure}(table, ..., platform = cloud_platform())

\S4method{avtables}{azure}(version = .WDS_API_VERSION, ..., platform = cloud_platform())

\S4method{avtable_import_set}{azure}(
  .data,
  origin,
  set = names(.data)[[1]],
  member = names(.data)[[2]],
  ...,
  platform = cloud_platform()
)
}
\arguments{
\item{table}{\code{character(1)} The name of the table / type}

\item{.data}{\code{tibble()} The dataset chiefly from the \code{avtable()} operation}

\item{origin}{\code{character(1)} name of the type (table) used to create the set
e.g "sample", "participant", etc.}

\item{set}{\code{character(1)} column name of \code{.data} identifying the set(s) to be
created.}

\item{member}{\code{character()} vector of entity from the \code{avtable} identified by
\code{origin}. The values may repeat if an ID is in more than one set}
}
\value{
\code{avtable_import_set()} returns a \code{character(1)} name of the imported
tibble.
}
\description{
Methods for working with AnVIL Azure tables. These are referred
to as "types" in the AnVIL Workspace Data Service (WDS) API.
}
\details{
\code{avtable_import_set()} creates new rows in a table \verb{<origin>_set}.
One row will be created for each distinct value in the column identified by
\code{set}. Each row entry has a corresponding column \verb{<origin>} linking to one
or more rows in the \verb{<origin>} table, as given in the \code{member} column. The
operation is somewhat like \code{split(member, set)}.
}
\section{Functions}{
\itemize{
\item \code{avtable(azure)}: List the contents of a particular table / type

\item \code{avtables(azure)}: List the available tables / types

\item \code{avtable_import_set(azure)}: Create a grouping table from an origin dataset

}}
\examples{
\dontrun{
avtable("testData") |>                          # new 'testData_set' table
    avtable_import_set("testData", "cyl", "model_id")
}
}
