% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avworkspace-methods.R
\name{avworkspace-methods}
\alias{avworkspace-methods}
\alias{avworkspaces,azure-method}
\alias{avworkspace_namespace,azure-method}
\alias{avworkspace_name,azure-method}
\alias{avworkspace,azure-method}
\title{AnVIL Azure Workspace methods}
\usage{
\S4method{avworkspaces}{azure}(..., platform = cloud_platform())

\S4method{avworkspace_namespace}{azure}(..., platform = cloud_platform())

\S4method{avworkspace_name}{azure}(..., platform = cloud_platform())

\S4method{avworkspace}{azure}(..., platform = cloud_platform())
}
\arguments{
\item{...}{Additional arguments passed to the underlying methods (not used).}

\item{platform}{\code{character(1)} The platform class to dispatch on. The default
is \code{"azure"} for the \code{AnVILAz} package.}
}
\description{
AnVIL Azure Workspace methods
}
\section{Functions}{
\itemize{
\item \code{avworkspaces(azure)}: List workspaces

\item \code{avworkspace_namespace(azure)}: List the workspace namespace

\item \code{avworkspace_name(azure)}: Obtain the workspace name

\item \code{avworkspace(azure)}: Obtain the current workspace namespace and
name combination

}}
