% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az_copy.R
\name{az_copy}
\alias{az_copy}
\alias{az_copy_copy}
\alias{az_copy_list}
\alias{az_copy_rm}
\title{Azure Copy command line utility interface}
\usage{
az_copy_copy(from, to)

az_copy_list()

az_copy_rm(blob_file)
}
\arguments{
\item{from}{\code{character(1)} A file path relevant to the Azure workspace,
usually where the Jupyter Notebooks are found.}

\item{to}{\code{character(1)} A relative path in the Azure Storage Container}

\item{blob_file}{\code{character(1)} A relative path to a file in the Azure
Storage Container to be removed}
}
\value{
\itemize{
\item \code{az_copy_list} - a \code{tibble} of files and metadata
\item \code{az_copy_copy} - called for the side effect of copying a file
\item \code{az_copy_rm} - called for the side effect of removing a file
}
}
\description{
These functions invoke the \code{azcopy} command line utility. The
utilities make use of a managed SAS token to mainly transfer files from the
Azure workspace to the Azure Storage container. See \code{get_sas_token} for
credential details.
}
\details{
\itemize{
\item \code{az_copy_list} - list all the files in the Azure Storage Container
\item \code{az_copy_copy} - copy a file from the Azure workspace to the container
\item \code{az_copy_rm} - remove a file from the Azure Storage Container
}
}
\examples{
if (interactive()) {

  az_copy_list()
  az_copy_copy("jupyter.log", "analyses/jupyter.log")
  az_copy_rm("analyses/jupyter.log")

}
}
