% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure-class.R
\docType{class}
\name{azure-class}
\alias{azure-class}
\alias{.azure}
\alias{azure}
\title{Azure platform class}
\usage{
azure()
}
\value{
An object of class \code{azure}.
}
\description{
This class represents the Azure platform.
}
\examples{
showClass("azure")

az <- azure()
az
showClass(class(az))
}
